Display '?' if there is a weird problem from `git rev-list --left-right`.
The '?' is only displayed when parsing `git rev-list --left-right` yields no upstream differences and the exit code from that command is nonzero. In other words, the command failed without any differences.

This failure mode includes the possibility when no upstream actually exists, so in passing, this change happens to provide a workable solution for #30.