Add config setting for ahead/behind display format
Setting `bash.branchBehindAndAheadDisplay`.

If `full`, displays the behind count and ahead count separately with the
appropriate arrows.

If `compact`, dispays the behind count, then the double arrow, then the
ahead count.

If `minimal` (or any other value), displays no counts or symbols. The
color of the branch name will still change according to the ahead/behind
status. This is consistent with dahlbyk/posh-git as of commit 4d8e08c on
2017-10-29; prior to that, `minimal` would have shown the arrows but no
counts.