Show local status summary symbol (! or ~)
Adds a `!` at the end of the local file status indicators if the repo
has unstaged changes, or a `~` if it has uncommitted (staged) changes
but no unstaged changes.

To achieve greater consistency with upstream posh-git, neither this
symbol nor the stash indicator will appear if `bash.enableFileStatus` is
`false`. Further, the rebase indicator has been moved to _before_ the
file status indicators, and is now immediately adjacent to the branch
name/status. This commit does _not_ change the color of the rebase
indicator, which in posh-git is cyan but in posh-git-sh is currently the
default terminal foreground color.

Resolves #40.